+++ modules/apis/src/test/java/com/axway/apim/cli/APIImportCLIOptionsTest.java
@@ -35,7 +35,7 @@ public class APIImportCLIOptionsTest {
 	
 	@Test
 	public void testAPIImportParameter() throws ParseException, AppException {
-		String[] args = {"-s", "prod", "-c", "myConfig.json", "-clientOrgsMode", "replace", "-clientAppsMode", "replace", "-quotaMode", "replace", "-detailsExportFile", "myExportFile.txt", "-stageConfig", "myStageConfigFile.json"};
+		String[] args = {"-s", "prod", "-c", "myConfig.json", "-clientOrgsMode", "replace", "-clientAppsMode", "replace", "-quotaMode", "replace", "-detailsExportFile", "myExportFile.txt", "-stageConfig", "myStageConfigFile.json", "-enabledCaches", "applicationsQuotaCache,*API*"};
 		CLIOptions options = CLIAPIImportOptions.create(args);
 		APIImportParams params = (APIImportParams) options.getParams();
 		Assert.assertEquals(params.getUsername(), "apiadmin");
@@ -46,6 +46,7 @@ public class APIImportCLIOptionsTest {
 		Assert.assertEquals(params.getQuotaMode(), Mode.replace);
 		Assert.assertEquals(params.getDetailsExportFile(), "myExportFile.txt");
 		Assert.assertEquals(params.getStageConfig(), "myStageConfigFile.json");
+		Assert.assertEquals(params.getEnabledCaches(), "applicationsQuotaCache,*API*");
 	}
 	
 	@Test