+++ modules/apps/src/main/java/com/axway/apim/appimport/lib/AppImportCLIOptions.java
@@ -5,6 +5,7 @@ import org.apache.commons.cli.Option;
 import com.axway.apim.lib.CLIOptions;
 import com.axway.apim.lib.CoreCLIOptions;
 import com.axway.apim.lib.Parameters;
+import com.axway.apim.lib.StandardImportCLIOptions;
 import com.axway.apim.lib.errorHandling.AppException;
 
 public class AppImportCLIOptions extends CLIOptions {
@@ -15,6 +16,7 @@ public class AppImportCLIOptions extends CLIOptions {
 	
 	public static CLIOptions create(String[] args) {
 		CLIOptions cliOptions = new AppImportCLIOptions(args);
+		cliOptions = new StandardImportCLIOptions(cliOptions);
 		cliOptions = new CoreCLIOptions(cliOptions);
 		cliOptions.addOptions();
 		cliOptions.parse();
@@ -27,10 +29,6 @@ public class AppImportCLIOptions extends CLIOptions {
 		Option option = new Option("c", "config", true, "This is the JSON-Formatted Application-Config file containing the application. You may get that config file using apim app get with output set to JSON.");
 		option.setRequired(true);
 		option.setArgName("app_config.json");
-		addOption(option);
-		
-		option = new Option("stageConfig", true, "Manually provide the name of the stage configuration file to use instead of derived from the given stage.");
-		option.setArgName("my-staged-app-config.json");
 		addOption(option);
 	}
 
@@ -59,7 +57,6 @@ public class AppImportCLIOptions extends CLIOptions {
 		AppImportParams params = new AppImportParams();
 		
 		params.setConfig(getValue("c"));
-		params.setStageConfig(getValue("stageConfig"));
 		return params;
 	}
 }