+++ modules/apps/src/main/java/com/axway/apim/appimport/lib/AppImportParams.java
@@ -8,10 +8,4 @@ public class AppImportParams extends StandardImportParams {
 	public static synchronized AppImportParams getInstance() {
 		return (AppImportParams)CoreParameters.getInstance();
 	}
-	
-	@Override
-	public boolean isIgnoreCache() {
-		// For import action we ignore the cache in all cases!
-		return true;
-	}
 }