+++ modules/organizations/src/main/java/com/axway/apim/organization/lib/OrgImportCLIOptions.java
@@ -5,6 +5,7 @@ import org.apache.commons.cli.Option;
 import com.axway.apim.lib.CLIOptions;
 import com.axway.apim.lib.CoreCLIOptions;
 import com.axway.apim.lib.Parameters;
+import com.axway.apim.lib.StandardImportCLIOptions;
 import com.axway.apim.lib.errorHandling.AppException;
 
 public class OrgImportCLIOptions extends CLIOptions {
@@ -15,6 +16,7 @@ public class OrgImportCLIOptions extends CLIOptions {
 	
 	public static CLIOptions create(String[] args) {
 		CLIOptions cliOptions = new OrgImportCLIOptions(args);
+		cliOptions = new StandardImportCLIOptions(cliOptions);
 		cliOptions = new CoreCLIOptions(cliOptions);
 		cliOptions.addOptions();
 		cliOptions.parse();
@@ -27,10 +29,6 @@ public class OrgImportCLIOptions extends CLIOptions {
 		Option option = new Option("c", "config", true, "This is the JSON-Formatted Organization-Config file containing the organization. You may get that config file using apim org get with output set to JSON.");
 		option.setRequired(true);
 		option.setArgName("org_config.json");
-		addOption(option);
-		
-		option = new Option("stageConfig", true, "Manually provide the name of the stage configuration file to use instead of derived from the given stage.");
-		option.setArgName("my-staged-org-config.json");
 		addOption(option);
 	}
 
@@ -57,7 +55,6 @@ public class OrgImportCLIOptions extends CLIOptions {
 	public Parameters getParams() throws AppException {
 		OrgImportParams params = new OrgImportParams();
 		params.setConfig(getValue("config"));
-		params.setStageConfig(getValue("stageConfig"));
 		return params;
 	}
 }