+++ modules/settings/src/main/java/com/axway/apim/setup/APIManagerSettingsApp.java
@@ -176,6 +176,10 @@ public class APIManagerSettingsApp implements APIMCLIServiceProvider {
 			LOG.error(e.getMessage(), e);
 			result.setError(ErrorCode.UNXPECTED_ERROR);
 			return result;
+		} finally {
+			try {
+				APIManagerAdapter.deleteInstance();
+			} catch (AppException ignore) { }
 		}
 	}